requirements-core.txt
streamlit
opencv-python
numpy
pandas
psycopg2-binary
email-validator
pygame
mediapipe
pyarrow>=7.0,<10.0
